refactor(suite): import labeling components directly after barrel removal
Update imports of AccountLabeling and WalletLabeling to point at their
component files now that the labeling/index.tsx barrel and the
components/suite re-exports have been removed.
